+++ packages/fluentui/react-context-selector/src/createContext.ts
@@ -1,33 +1,38 @@
 import * as React from 'react';
-import { Context, ContextListener, ContextValue, CreateContextOptions } from './types';
 
-// Stops React Context propagation
-// https://github.com/facebook/react/blob/95bd7aad7daa80c381faa3215c80b0906ab5ead5/packages/react-reconciler/src/ReactFiberBeginWork.js#L2656
-const calculateChangedBits = () => 0;
+import { Context, ContextValue } from './types';
+import { runWithNormalPriority, useIsomorphicLayoutEffect } from './utils';
 
 const createProvider = <Value>(Original: React.Provider<ContextValue<Value>>) => {
   const Provider: React.FC<React.ProviderProps<Value>> = props => {
-    const listeners = React.useRef<ContextListener<Value>[]>([]);
+    // Holds an actual "props.value"
+    const valueRef = React.useRef(props.value);
+    // Used to sync context updates and avoid stale values, can be considered as render/effect counter of Provider.
+    const versionRef = React.useRef(0);
 
-    // We call listeners in render intentionally. Listeners are not technically pure, but
-    // otherwise we can't get benefits from concurrent mode.
-    //
-    // We make sure to work with double or more invocation of listeners.
-    listeners.current.forEach(listener => listener(props.value));
+    // A stable object, is used to avoid context updates via mutation of its values.
+    const contextValue = React.useRef<ContextValue<Value>>();
 
-    // Disables updates propogation for React Context as `value` is always shallow equal
-    const subscribe = React.useCallback((listener: ContextListener<Value>) => {
-      listeners.current.push(listener);
-
-      const unsubscribe = () => {
-        const index = listeners.current.indexOf(listener);
-        listeners.current.splice(index, 1);
+    if (!contextValue.current) {
+      contextValue.current = {
+        value: valueRef,
+        version: versionRef,
+        listeners: [],
       };
+    }
 
-      return unsubscribe;
-    }, []);
+    useIsomorphicLayoutEffect(() => {
+      valueRef.current = props.value;
+      versionRef.current += 1;
 
-    return React.createElement(Original, { value: { subscribe, value: props.value } }, props.children);
+      runWithNormalPriority(() => {
+        (contextValue.current as ContextValue<Value>).listeners.forEach(listener => {
+          listener([versionRef.current, props.value]);
+        });
+      });
+    }, [props.value]);
+
+    return React.createElement(Original, { value: contextValue.current }, props.children);
   };
 
   /* istanbul ignore else */
@@ -38,28 +43,13 @@ const createProvider = <Value>(Original: React.Provider<ContextValue<Value>>) =>
   return Provider;
 };
 
-export const createContext = <Value>(defaultValue: Value, options: CreateContextOptions = {}): Context<Value> => {
-  const { strict = true } = options;
+export const createContext = <Value>(defaultValue: Value): Context<Value> => {
+  const context = React.createContext<ContextValue<Value>>({
+    value: { current: defaultValue },
+    version: { current: -1 },
+    listeners: [],
+  });
 
-  const context = React.createContext<ContextValue<Value>>(
-    {
-      get subscribe() {
-        if (strict) {
-          /* istanbul ignore next */
-          throw new Error(
-            process.env.NODE_ENV === 'production'
-              ? ''
-              : `Please use <Provider /> component from "@fluentui/react-context-selector"`,
-          );
-        }
-
-        /* istanbul ignore next */
-        return () => () => {};
-      },
-      value: defaultValue,
-    },
-    calculateChangedBits,
-  );
   context.Provider = createProvider<Value>(context.Provider) as any;
 
   // We don't support Consumer API